final commit, adds server app
