--- package.json
@@ -1,20 +1,0 @@
-{
-  "name": "gitlab-cli",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "body-parser": "^1.20.0",
-    "cors": "^2.8.5",
-    "esbuild": "^0.12.21",
-    "express": "^4.17.1",
-    "p-queue": "^7.2.0",
-    "simple-git": "^3.7.1"
-  },
-  "scripts": {
-    "build": "rm -rf dist && node build.js"
-  },
-  "devDependencies": {
-    "prettier": "^2.3.2"
-  }
-}
