+++ server-app/main.js
@@ -1,0 +1,28 @@
+const electron = require("electron");
+const path = require("path");
+
+const { app, BrowserWindow } = electron;
+
+const createWindow = () => {
+  const win = new BrowserWindow({
+    width: 1080,
+    height: 768,
+    webPreferences: {
+      nodeIntegration: true,
+      contextIsolation: false,
+      enableRemoteModule: true,
+    },
+  });
+
+  win.loadFile(path.join(__dirname, "./views/index.html"));
+};
+
+app.whenReady().then(() => {
+  createWindow();
+
+  app.on("activate", () => {
+    if (BrowserWindow.getAllWindows().length === 0) {
+      createWindow();
+    }
+  });
+});
