+++ server-app/server/app.js
@@ -1,78 +1,36 @@
-"use strict";
-
 const express = require("express");
-const git = require("simple-git");
-const bodyParser = require("body-parser");
 const cors = require("cors");
-
-import PQueue from "p-queue";
-import { getLocalRepository } from "./utils";
-import { mergeProcess } from "./merge";
-import { getMergeCommits } from "./getMergeCommits";
-import { cherryPickProcess } from "./cherryPick";
+const bodyParser = require("body-parser");
+const PQueue = require("p-queue");
+import { getLocalRepository } from "./utils/helper.js";
+import { mergeProcess } from "./utils/merge.js";
+import { getMergeCommits } from "./utils/getMergeCommits.js";
+import { cherryPickProcess } from "./utils/cherryPick.js";
 import {
   getProfiles,
   addProfile,
   deleteProfile,
   updateProfile,
-} from "./profiles";
+} from "./utils/profiles.js";
+import { log, init as initLogger } from "./logger.js";
 
-const queue = new PQueue({ concurrency: 1 });
-
+//Globals
 const PORT = 4000;
+
 const app = express();
+app.use(cors());
 app.use(bodyParser.json());
 app.use(
   bodyParser.urlencoded({
     extended: true,
   })
 );
-app.use(cors());
 
-app.get("/handshake", async function (req, res) {
-  try {
-    console.log("Handshake successful !!");
-    res.status(200).end();
-    res.end();
-  } catch (e) {
-    res.status(400).end();
-  }
-});
+const queue = new PQueue({ concurrency: 1 });
 
-app.get("/profiles", async function (req, res) {
-  try {
-    const profileResponse = await getProfiles();
-    res.status(200).send(profileResponse);
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-app.post("/profiles", async function (req, res) {
-  try {
-    const profileResponse = await addProfile(req.body);
-    res.status(200).send(profileResponse);
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-app.delete("/profiles", async function (req, res) {
-  try {
-    const profileResponse = await deleteProfile(req.body.id);
-    res.status(200).send(profileResponse);
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-app.put("/profiles", async function (req, res) {
-  try {
-    const profileResponse = await updateProfile(
-      req.body.id,
-      req.body.profileData
-    );
-    res.status(200).send(profileResponse);
-  } catch (e) {
-    res.status(400).send(e);
-  }
+//Routes
+app.get("/", (req, res) => {
+  res.send("App is running");
 });
 
 app.get("/merge", async function (req, res) {
@@ -84,12 +42,64 @@ app.get("/merge", async function (req, res) {
   try {
     const { source, target, location } = req.query;
     const localRepo = await getLocalRepository(location);
-    res.write(`Merge Queued `);
+    res.write(`Merge Queued`);
+    log(`Merge Queued`);
     queue.add(
       async () => await mergeProcess(res, source, target, localRepo.path)
     );
   } catch (e) {
     res.write(e.toString());
+  }
+});
+
+app.get("/handshake", async function (req, res) {
+  try {
+    res.status(200).end();
+    log("Handshake successful !!");
+    res.end();
+  } catch (e) {
+    log(e);
+    res.status(400).end();
+  }
+});
+
+app.get("/profiles", async function (req, res) {
+  try {
+    const profileResponse = await getProfiles();
+    res.status(200).send(profileResponse);
+  } catch (e) {
+    log(e);
+    res.status(400).send(e);
+  }
+});
+app.post("/profiles", async function (req, res) {
+  try {
+    const profileResponse = await addProfile(req.body);
+    res.status(200).send(profileResponse);
+  } catch (e) {
+    log(e);
+    res.status(400).send(e);
+  }
+});
+app.delete("/profiles", async function (req, res) {
+  try {
+    const profileResponse = await deleteProfile(req.body.id);
+    res.status(200).send(profileResponse);
+  } catch (e) {
+    log(e);
+    res.status(400).send(e);
+  }
+});
+app.put("/profiles", async function (req, res) {
+  try {
+    const profileResponse = await updateProfile(
+      req.body.id,
+      req.body.profileData
+    );
+    res.status(200).send(profileResponse);
+  } catch (e) {
+    log(e);
+    res.status(400).send(e);
   }
 });
 
@@ -100,6 +110,7 @@ app.post("/cherrypick", async function (req, res) {
     "access-control-allow-origin": "*",
   });
   res.write(`Cherry-Pick Queued `);
+  log(`Cherry-Pick Queued`);
   queue.add(async () => await cherryPickProcess(req, res));
 });
 
@@ -120,10 +131,19 @@ app.post("/mergecommits", async function (req, res) {
     });
     res.end(JSON.stringify(jsonResponse));
   } catch (e) {
+    log(e);
     res.status(400).end(e.toString());
   }
 });
 
-app.listen(PORT, () => {
-  console.log("Gitlab CLI listening at 4000...");
-});
+const init = async () => {
+  try{
+    await initLogger();
+    app.listen(PORT);
+  }
+  catch(e){
+    log(e);
+  }
+};
+
+init();
