+++ server-app/server/utils/helper.js
@@ -7,7 +7,7 @@ export function wait(millis) {
 
 export async function readConfigFile() {
   try {
-    const configPath = path.join(__dirname, "config.json");
+    const configPath = path.join(__dirname, '../server/resources/config.json');
     const configData = await fs.readFile(configPath);
     return JSON.parse(configData);
   } catch (e) {
@@ -17,7 +17,7 @@ export async function readConfigFile() {
 
 export async function writeConfigFile(configJSONData) {
   try {
-    const configPath = path.join(__dirname, "config.json");
+    const configPath = path.join(__dirname, '../server/resources/config.json');
     await fs.writeFile(configPath, JSON.stringify(configJSONData));
   } catch (e) {
     throw new Error(e.toString());
@@ -51,4 +51,8 @@ export function renderPauseMessage(currentCommitSHA, e) {
     </div>
   </div>
   <div class="cherrypick-error-desc">${e.toString()}</div>`;
+}
+
+export function getLogFilePath() {
+    return path.join(__dirname, '../server/resources/console.txt');
 }
