+++ server-app/views/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
+    <title>GitLab CLI</title>
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
       rel="stylesheet"
@@ -8,33 +9,32 @@
       crossorigin="anonymous"
     />
     <link rel="stylesheet" href="style.css" />
-    <script src="assets/profileScript.js" defer></script>
+    <script type="module" src="../server/app.js"></script>
+    <script type="module" src="./index.js"></script>
   </head>
   <body>
     <div class="app-header">
-      <h2>GitLab CLI Extension</h2>
+      <h2>GitLab CLI</h2>
     </div>
     <form class="add-profile-form">
       <div class="form-group row">
         <div class="form-group col-md-6">
-          <label for="url" class="required">Repository URL</label>
+          <label for="url">Repository URL</label>
           <input
             type="text"
             class="form-control"
             id="url"
             name="url"
-            placeholder="Homepage url of the gitLab repo"
             required
           />
         </div>
         <div class="form-group col-md-6">
-          <label for="path" class="required">Local Path</label>
+          <label for="path">Local Path</label>
           <input
             type="text"
             class="form-control"
             id="path"
             name="path"
-            placeholder="absolute path of the local repo"
             required
           />
         </div>
@@ -45,5 +45,10 @@
       </button>
     </form>
     <div id="options-desc"></div>
+    <div id="console-box">
+      <button id="console-expand">Expand</button>
+      Server started at Port 4000
+      <div id="console-desc"></div>
+    </div>
   </body>
 </html>
