+++ server-app/views/style.css
@@ -16,7 +16,9 @@ input {
 }
 form {
   margin: auto;
-  width: 70%;
+  width: 85%;
+  max-height: 200px;
+  overflow-y: scroll;
   border: 1px solid #e5e5e5;
   margin-top: 30px;
   padding: 15px;
@@ -31,10 +33,11 @@ h2 {
   text-align: center;
   color: white;
 }
-#cherry-pick-desc, #options-desc {
+#cherry-pick-desc,
+#options-desc {
   display: none;
   margin: auto;
-  width: 70%;
+  width: 85%;
   padding: 8px;
   margin-top: 5px;
   border: 1px solid #e5e5e5;
@@ -42,24 +45,37 @@ h2 {
   border-radius: 4px;
   box-sizing: border-box;
   font-size: 15px;
-  word-break: break-all;
 }
-.copy-button{
+#console-box {
+  margin: auto;
+  width: 100%;
+  height: 200px;
+  overflow: scroll;
+  overflow-y: scroll;
+  word-break: break-all;
+  padding: 8px;
+  border: 1px solid #e5e5e5;
+  color: white;
+  font-family: monospace;
+  background-color: #000000;
+  border-radius: 4px;
+  box-sizing: border-box;
+  font-size: 15px;
+  position: fixed;
+  bottom: 0;
+}
+.copy-button {
   border: none;
   padding: 0;
   background: none;
 }
-.cherrypick-error-desc{
-  max-height: 150px;
-  word-break: break-all;
-  overflow-y: scroll;
-}
-.card{
-  margin-top: 8px;
-  margin-bottom: 8px;
-}
-.required:after {
-  color: #e32;
-  content: ' *';
-  display:inline;
+#console-expand {
+  padding: "2px";
+  overflow: scroll;
+  padding-left: "4px";
+  padding-right: "4px";
+  border: 1px solid rgb(0, 225, 255);
+  background: black;
+  float: right;
+  color: rgb(0, 225, 255);
 }
