--- src/extension/constants/domClasses.ts
@@ -1,4 +1,0 @@
-export const DETAIL_PAGE_DESCRIPTION = '.detail-page-description';
-export const GITLAB_CLI_DESC= 'gitlab-cli-desc';
-export const GITLAB_CLI_BUTTON='js-gitlab-cli-button';
-export const MR_WIDGET_SECTION='.mr-widget-section';
