--- src/extension/log.ts
@@ -1,3 +1,0 @@
-export function log(...args: string[]) {
-  return console.log('Gitlab CLI Merge: ', ...args);
-}
