--- src/extension/public/cherrypick.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-      crossorigin="anonymous"
-    />
-    <link rel="stylesheet" href="style.css" />
-    <script src="assets/cherryPickForm.js" defer></script>
-  </head>
-  <body>
-    <div class="app-header">
-      <h2>GitLab CLI Extension</h2>
-    </div>
-    <form class="cherry-pick-form">
-      <div class="form-group row">
-        <div class="form-group col-md-6">
-          <label for="location" class="required">Repository URL:</label>
-          <input
-            type="text"
-            class="form-control"
-            autocomplete="off"
-            id="location"
-            name="location"
-            list="url-list"
-            placeholder="select repo url present in config file"
-            required
-          />
-          <datalist id="url-list"></datalist>
-        </div>
-        <div class="form-group col-md-6">
-          <label for="fetchMergeCommitBranch">Fetch Merge Commits From:</label>
-          <input
-            type="text"
-            class="form-control"
-            id="fetchMergeCommitBranch"
-            name="fetchMergeCommitBranch"
-            placeholder="branch to get merge commits from (leave blank to get all merge commits)"
-          />
-        </div>
-      </div>
-      <div class="form-group row">
-        <div class="form-group col-md-6">
-          <label for="commitBranch" class="required">Source Branch</label>
-          <input
-            type="text"
-            class="form-control"
-            id="commitBranch"
-            name="commitBranch"
-            placeholder="new branch or existing branch in origin"
-            required
-          />
-        </div>
-        <div class="form-group col-md-6">
-          <label for="targetBranch" class="required">Target Branch</label>
-          <input
-            type="text"
-            class="form-control"
-            id="targetBranch"
-            name="targetBranch"
-            placeholder="target branch"
-            required
-          />
-        </div>
-      </div>
-      <div class="form-group row">
-        <div class="form-group col-md-6">
-          <label for="commitAuthor" class="required">Author</label>
-          <input
-            type="text"
-            class="form-control"
-            id="commitAuthor"
-            name="commitAuthor"
-            placeholder="email id of author for whom to fetch merge commits for"
-            required
-          />
-        </div>
-        <div class="form-group col-md-6">
-          <label for="commitTime" class="required">Time since commit:</label>
-          <input
-            type="datetime-local"
-            class="form-control"
-            id="commitTime"
-            name="commitTime"
-            required
-          />
-        </div>
-      </div>
-      <button type="submit" class="btn btn-primary">Submit</button>
-    </form>
-    <div id="cherry-pick-desc"></div>
-  </body>
-</html>
