--- src/extension/public/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-      crossorigin="anonymous"
-    />
-    <script src="assets/popup.js" defer></script>
-    <link rel="stylesheet" type="text/css" href="./indexStyle.css" />
-  </head>
-  <body>
-    <div class="app-header">
-      <h2>GitLab CLI Extension</h2>
-    </div>
-    <div style="display: flex;justify-content: space-between;">
-    <button class="btn btn-primary" id="btn-cherry-pick">Cherry pick</button>
-    <button class="btn btn-primary" id="btn-options">Profiles</button>
-  </div>
-  </body>
-</html>
