--- src/server/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "gitlab-cli-server",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "scripts": {
-    "start": "node ./api.js"
-  }
-}
